Correct typo in the build configuration of GitHub workflow
Corrected a typo in the build configuration of the GitHub workflow from 'desc' to 'dest' in line 19. This was misnaming the destination file for the zipping process, impeding the correct the creation of 'build.zip' from the files in the 'out' directory. This fix will ensure that the workflow runs as expected and produces a 'build.zip' for the archive upload.